+++ test/test_helper.exs
@@ -1,1 +1,2 @@
 ExUnit.start()
+Code.require_file("test/support/generators.ex")